cloud_functions/instance_stop/requirements.txt
functions-framework>=3.0.0
cloudevents>=1.10.0
google-api-core
google-cloud-compute